feat: Add load_data function to read data from SQLite database
This function adds the capability to load data from the DisasterResponse.db SQLite database, specifically from the 'messages' table. It uses a SQL query to select all columns from the table and returns the result as a Pandas DataFrame. This DataFrame will be used as input for the machine learning model training process.